Extract language review summary to localization directory
